requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
aiosqlite==0.17.0
anyio==3.6.1
asgiref==3.5.2
async-timeout==4.0.2
atlastk==0.13
attrs==21.4.0
base58==2.1.1
bitarray==2.5.1
certifi==2022.6.15
charset-normalizer==2.0.12
click==8.1.3
cytoolz==0.11.2
eth-abi==2.1.1
eth-account==0.5.8
eth-hash==0.3.2
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.10.0
fastapi==0.78.0
frozenlist==1.3.0
h11==0.13.0
hexbytes==0.2.2
idna==3.3
importlib-resources==5.8.0
install==1.3.5
ipfshttpclient==0.8.0a2
iso8601==1.0.2
jsonschema==4.6.0
lru-dict==1.1.7
marshmallow==3.16.0
marshmallow-dataclass==8.5.8
multiaddr==0.0.9
multidict==6.0.2
mypy-extensions==0.4.3
netaddr==0.8.0
packaging==21.3
parsimonious==0.8.1
protobuf==3.20.1
pycryptodome==3.14.1
pydantic==1.9.1
pyparsing==3.0.9
pypika-tortoise==0.1.5
pyrsistent==0.18.1
pytz==2022.1
PyYAML==6.0
requests==2.28.0
rlp==2.0.1
six==1.16.0
sniffio==1.2.0
starlette==0.19.1
toolz==0.11.2
tortoise==0.1.1
tortoise-orm==0.19.1
typing-inspect==0.7.1
typing_extensions==4.2.0
urllib3==1.26.9
uvicorn==0.17.6
varint==1.0.2
web3==5.29.2
websockets==9.1
yarl==1.7.2
zipp==3.8.0
